+++ core/deezer_client.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional, Any
 from functools import wraps
 from dataclasses import dataclass
 from utils.logging_config import get_logger
-from core.metadata_cache import get_metadata_cache
+from core.metadata.cache import get_metadata_cache
 
 logger = get_logger("deezer_client")
