+++ core/metadata/cache.py
